refactor: PrezelBadge 컴포넌트 구조 개선 및 속성 타입 변경
`PrezelBadge` 컴포넌트의 위치를 이동하고, 불필요한 Enum 타입을 Boolean으로 대체하여 구조를 단순화했습니다.

*   `PrezelBadge.kt` 위치 변경: `component/badge/` -> `component/`
*   `PrezelBadgeActive`, `PrezelBadgeDisabled` Enum 삭제 및 `Boolean` 타입으로 변경
*   `PrezelBadgeSize` Enum을 `PrezelBadge.kt` 파일 내부로 이동 및 `PrezelBadgeSize.kt` 삭제
*   패키지 경로 변경에 따른 import 문 수정 및 프리뷰 코드 업데이트
