--- Prezel/core/designsystem/src/main/java/com/team/prezel/core/designsystem/component/badge/PrezelBadgeSize.kt
@@ -1,16 +1,0 @@
-package com.team.prezel.core.designsystem.component.badge
-
-enum class PrezelBadgeSize {
-    SMALL,
-    REGULAR,
-}
-
-enum class PrezelBadgeActive {
-    ON,
-    OFF,
-}
-
-enum class PrezelBadgeDisabled {
-    ON,
-    OFF,
-}
